Fix denort code and lint
